Move WebAssemblyBlazorApplicationBuilder
