+++ src/Components/blazor/src/Microsoft.AspNetCore.Blazor/Hosting/WebAssemblyBlazorApplicationBuilder.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Blazor.Rendering;
 using Microsoft.AspNetCore.Components.Builder;
 
-namespace Microsoft.AspNetCore.Blazor.Builder
+namespace Microsoft.AspNetCore.Blazor.Hosting
 {
     internal class WebAssemblyBlazorApplicationBuilder : IBlazorApplicationBuilder
     {
